src/agents/new-agents/gmail_agent/requirements.txt
uagents>=0.18.1
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
requests>=2.25.0
openai>=1.0.0